requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Faker==37.8.0
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pefile==2023.2.7
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==80.9.0
tzdata==2025.2
Werkzeug==3.1.3
